FBSD: fix for use-after-free problem in 10.x in sys/kern/subr_taskqueue.c
In 11 and 12 the taskqueue code has been rewritten in this area but
under 10 this bug still occurs.

On our appliances this bug stops the system from mounting the ZFS
root, so it is quite severe.
Basically while the thread is sleeping during the ZFS mount of root
(in the while loop), another thread can free the 'task' item it is
checking in that while loop and it can be reused or filled with
'deadcode' etc., with the waiting code unaware of the change.. The
fix is to refetch the item at the end of the queue each time around
the loop.
I don't really want to do the bigger change of MFCing the change in
11, as it is more extensive, though if someone else does, that's ok
by me. (If it's ABI compatible)

Patch-by: Julian Elischer <julian@freebsd.org>
Message-ID: <7b732876-8cc3-a638-7ff1-e664060d4907@freebsd.org>